Add integration tests for add, remove Item and set quantity for Item to ShoppingList REST controller test
